!perf: Use string for iteration and error reporting
The string type is smaller than a byte slice and can be
copied much faster than []byte on modern 64-bit systems.
Drawbacks: this change requires the use of unsafe.
Type Error is no longer required to be type-parameterized but
it's preserved as Error[S] in order to remain backward compatible.
The type parameter can be removed in a new major version.